Remove handle_command() to command.cpp
