+++ client/include/command.h
@@ -1,0 +1,8 @@
+#ifndef COMMAND_H
+#define COMMAND_H
+
+#include <string>
+
+int handle_command(int socket, const std::string& command);
+
+#endif /*COMMAND_H*/
